requirements.txt
beautifulsoup4==4.11.1
Flask==2.2.2
Jinja2==3.1.2
requests==2.28.1
discord.py==2.3.1